+++ packages/components/src/components/common/no-results-option/no-results-option-styles.spec.ts
@@ -4,10 +4,10 @@ import { getFunctionalComponentNoResultsOptionStyles } from './no-results-option
 
 describe('getFunctionalComponentNoResultsOptionStyles()', () => {
   it.each<Parameters<typeof getFunctionalComponentNoResultsOptionStyles>>([
-    ['select-option', `var(--p-internal-select-scaling, 1)`],
-    ['select-option', `var(--p-internal-select-scaling, 1)`],
-    ['multi-select-option', `var(--p-internal-multi-select-scaling, 1)`],
-    ['multi-select-option', `var(--p-internal-multi-select-scaling, 1)`],
+    ['select-option', `var(--_p-select-a, 1)`],
+    ['select-option', `var(--_p-select-a, 1)`],
+    ['multi-select-option', `var(--_p-multi-select-a, 1)`],
+    ['multi-select-option', `var(--_p-multi-select-a, 1)`],
   ])('should return correct css for componentName: %s and cssVarScaling: %s', (...args) => {
     validateCssAndMatchSnapshot(getCss(getFunctionalComponentNoResultsOptionStyles(...args)));
   });