+++ packages/components/src/components/table/table-cell/table-cell-styles.ts
@@ -1,5 +1,4 @@
 import { addImportantToEachRule, hostHiddenStyles } from '../../../styles';
-import { spacingFluidSm } from '../../../styles/css-variables';
 import { getCss } from '../../../utils';
 import { cssVariableTablePadding } from '../table/table-styles';
 
@@ -10,7 +9,7 @@ export const getComponentCss = (multiline: boolean): string => {
         display: 'table-cell',
         verticalAlign: 'middle',
         ...addImportantToEachRule({
-          padding: `var(${cssVariableTablePadding}, ${spacingFluidSm})`,
+          padding: `var(${cssVariableTablePadding})`,
           margin: 0,
           whiteSpace: multiline ? 'normal' : 'nowrap',
           ...hostHiddenStyles,