+++ packages/tokens/src/color/light/semantic/colorInfoFrostedLight.ts
@@ -1,3 +1,1 @@
-import { palette } from '../../palette';
-
-export const colorInfoFrostedLight = palette.light.blue['100a'];
+export const colorInfoFrostedLight = 'hsl(211 100% 90% / 0.55)';